+++ src/main/java/FIS/iLUVit/domain/iluvit/CenterApprovalReceivedAlarm.java
@@ -1,7 +1,6 @@
-package FIS.iLUVit.domain.iluvit.alarms;
+package FIS.iLUVit.domain.iluvit;
 
 import FIS.iLUVit.dto.alarm.AlarmDetailDto;
-import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.domain.common.Center;
 import FIS.iLUVit.domain.iluvit.enumtype.Auth;
 import FIS.iLUVit.service.AlarmUtils;
