+++ src/main/java/FIS/iLUVit/domain/iluvit/ConvertedToParticipateAlarm.java
@@ -1,8 +1,6 @@
-package FIS.iLUVit.domain.iluvit.alarms;
+package FIS.iLUVit.domain.iluvit;
 
 import FIS.iLUVit.domain.common.Center;
-import FIS.iLUVit.domain.iluvit.Presentation;
-import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.dto.alarm.AlarmDetailDto;
 import FIS.iLUVit.service.AlarmUtils;
 import lombok.Getter;
