+++ src/main/java/FIS/iLUVit/domain/iluvit/PresentationPeriodClosedAlarm.java
@@ -1,9 +1,7 @@
-package FIS.iLUVit.domain.iluvit.alarms;
+package FIS.iLUVit.domain.iluvit;
 
 import FIS.iLUVit.dto.alarm.AlarmDetailDto;
 import FIS.iLUVit.domain.common.Center;
-import FIS.iLUVit.domain.iluvit.Presentation;
-import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.service.AlarmUtils;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,7 +13,7 @@ import java.time.LocalDateTime;
 @Entity
 @NoArgsConstructor
 @Getter
-public class PresentationFullAlarm extends Alarm{
+public class PresentationPeriodClosedAlarm extends Alarm{
 
     @Column(name = "presentationId")
     private Long presentationId;
@@ -26,9 +24,9 @@ public class PresentationFullAlarm extends Alarm{
     @Column(name = "centerName")
     private String centerName;
 
-    public PresentationFullAlarm(User user, Presentation presentation, Center center) {
+    public PresentationPeriodClosedAlarm(User user, Presentation presentation, Center center) {
         super(user);
-        this.mode = AlarmUtils.PRESENTATION_APPLICANTS_FULL;
+        this.mode = AlarmUtils.PRESENTATION_CLOSED;
         this.presentationId = presentation.getId();
         this.centerId = center.getId();
         this.centerName = center.getName();
@@ -37,21 +35,22 @@ public class PresentationFullAlarm extends Alarm{
 
     @Override
     public AlarmDetailDto exportAlarm() {
-        return new PresentationFullAlarmDto(id, createdDate, message, dtype, presentationId, centerId, centerName);
+        return new PresentationPeriodClosedAlarmDto(id, createdDate, message, dtype, presentationId, centerId, centerName);
     }
 
     @Getter
-    public static class PresentationFullAlarmDto extends AlarmDetailDto {
+    public static class PresentationPeriodClosedAlarmDto extends AlarmDetailDto {
 
         protected Long presentationId;
         protected Long centerId;
         protected String centerName;
 
-        public PresentationFullAlarmDto(Long id, LocalDateTime createdDate, String message, String type, Long presentationId, Long centerId, String centerName) {
+        public PresentationPeriodClosedAlarmDto(Long id, LocalDateTime createdDate, String message, String type, Long presentationId, Long centerId, String centerName) {
             super(id, createdDate, message, type);
             this.presentationId = presentationId;
             this.centerId = centerId;
             this.centerName = centerName;
         }
     }
+
 }
