+++ src/main/java/FIS/iLUVit/domain/iluvit/ReportDetailPost.java
@@ -1,7 +1,5 @@
-package FIS.iLUVit.domain.iluvit.reports;
+package FIS.iLUVit.domain.iluvit;
 
-import FIS.iLUVit.domain.iluvit.Post;
-import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportReason;
 import lombok.Builder;
 import lombok.Getter;
