+++ src/main/java/FIS/iLUVit/event/AlarmEvent.java
@@ -1,6 +1,6 @@
 package FIS.iLUVit.event;
 
-import FIS.iLUVit.domain.iluvit.alarms.Alarm;
+import FIS.iLUVit.domain.iluvit.Alarm;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
