+++ src/main/java/FIS/iLUVit/event/eventListener/ParticipationCancelEventHandler.java
@@ -4,8 +4,8 @@ import FIS.iLUVit.domain.iluvit.Participation;
 import FIS.iLUVit.domain.iluvit.Presentation;
 import FIS.iLUVit.domain.iluvit.PtDate;
 import FIS.iLUVit.domain.iluvit.Waiting;
-import FIS.iLUVit.domain.iluvit.alarms.Alarm;
-import FIS.iLUVit.domain.iluvit.alarms.ConvertedToParticipateAlarm;
+import FIS.iLUVit.domain.iluvit.Alarm;
+import FIS.iLUVit.domain.iluvit.ConvertedToParticipateAlarm;
 import FIS.iLUVit.event.ParticipationCancelEvent;
 import FIS.iLUVit.repository.iluvit.AlarmRepository;
 import FIS.iLUVit.repository.iluvit.ParticipationRepository;
