+++ src/main/java/FIS/iLUVit/repository/iluvit/ReportRepository.java
@@ -1,6 +1,6 @@
 package FIS.iLUVit.repository.iluvit;
 
-import FIS.iLUVit.domain.iluvit.reports.Report;
+import FIS.iLUVit.domain.iluvit.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
