+++ src/main/java/FIS/iLUVit/service/ChatService.java
@@ -1,12 +1,12 @@
 package FIS.iLUVit.service;
 
-import FIS.iLUVit.domain.iluvit.alarms.Alarm;
+import FIS.iLUVit.domain.iluvit.Alarm;
 import FIS.iLUVit.domain.iluvit.*;
 import FIS.iLUVit.dto.chat.ChatDto;
 import FIS.iLUVit.dto.chat.ChatListDto;
 import FIS.iLUVit.dto.chat.ChatRequest;
 import FIS.iLUVit.dto.chat.ChatRoomRequest;
-import FIS.iLUVit.domain.iluvit.alarms.ChatAlarm;
+import FIS.iLUVit.domain.iluvit.ChatAlarm;
 import FIS.iLUVit.exception.*;
 import FIS.iLUVit.repository.iluvit.*;
 import lombok.RequiredArgsConstructor;
