+++ src/main/java/FIS/iLUVit/service/ReportService.java
@@ -5,9 +5,9 @@ import FIS.iLUVit.domain.iluvit.Comment;
 import FIS.iLUVit.domain.iluvit.Post;
 import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportType;
-import FIS.iLUVit.domain.iluvit.reports.Report;
-import FIS.iLUVit.domain.iluvit.reports.ReportDetailComment;
-import FIS.iLUVit.domain.iluvit.reports.ReportDetailPost;
+import FIS.iLUVit.domain.iluvit.Report;
+import FIS.iLUVit.domain.iluvit.ReportDetailComment;
+import FIS.iLUVit.domain.iluvit.ReportDetailPost;
 import FIS.iLUVit.exception.*;
 import FIS.iLUVit.repository.iluvit.*;
 import lombok.RequiredArgsConstructor;
