+++ src/test/java/FIS/iLUVit/controller/ReportControllerTest.java
@@ -9,8 +9,8 @@ import FIS.iLUVit.domain.iluvit.Teacher;
 import FIS.iLUVit.domain.iluvit.User;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportReason;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportType;
-import FIS.iLUVit.domain.iluvit.reports.Report;
-import FIS.iLUVit.domain.iluvit.reports.ReportDetail;
+import FIS.iLUVit.domain.iluvit.Report;
+import FIS.iLUVit.domain.iluvit.ReportDetail;
 import FIS.iLUVit.exception.*;
 import FIS.iLUVit.exception.exceptionHandler.ErrorResponse;
 import FIS.iLUVit.exception.exceptionHandler.controllerAdvice.GlobalControllerAdvice;
