+++ src/test/java/FIS/iLUVit/repository/ReportDetailRepositoryTest.java
@@ -7,8 +7,8 @@ import FIS.iLUVit.domain.iluvit.Post;
 import FIS.iLUVit.domain.iluvit.Teacher;
 import FIS.iLUVit.domain.iluvit.enumtype.Approval;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportType;
-import FIS.iLUVit.domain.iluvit.reports.Report;
-import FIS.iLUVit.domain.iluvit.reports.ReportDetail;
+import FIS.iLUVit.domain.iluvit.Report;
+import FIS.iLUVit.domain.iluvit.ReportDetail;
 import FIS.iLUVit.repository.iluvit.ReportDetailRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
