+++ src/test/java/FIS/iLUVit/service/ChatServiceTest.java
@@ -6,8 +6,8 @@ import FIS.iLUVit.dto.chat.ChatDto;
 import FIS.iLUVit.dto.chat.ChatListDto;
 import FIS.iLUVit.dto.chat.ChatRequest;
 import FIS.iLUVit.dto.chat.ChatRoomRequest;
-import FIS.iLUVit.domain.iluvit.alarms.Alarm;
-import FIS.iLUVit.domain.iluvit.alarms.ChatAlarm;
+import FIS.iLUVit.domain.iluvit.Alarm;
+import FIS.iLUVit.domain.iluvit.ChatAlarm;
 import FIS.iLUVit.domain.iluvit.enumtype.Auth;
 import FIS.iLUVit.domain.iluvit.enumtype.BoardKind;
 import FIS.iLUVit.event.AlarmEvent;
