+++ src/test/java/FIS/iLUVit/service/CommentServiceTest.java
@@ -4,8 +4,8 @@ import FIS.iLUVit.Creator;
 import FIS.iLUVit.domain.iluvit.*;
 import FIS.iLUVit.dto.comment.CommentDto;
 import FIS.iLUVit.dto.comment.CommentRequest;
-import FIS.iLUVit.domain.iluvit.alarms.Alarm;
-import FIS.iLUVit.domain.iluvit.alarms.PostAlarm;
+import FIS.iLUVit.domain.iluvit.Alarm;
+import FIS.iLUVit.domain.iluvit.PostAlarm;
 import FIS.iLUVit.domain.iluvit.enumtype.Auth;
 import FIS.iLUVit.domain.iluvit.enumtype.BoardKind;
 import FIS.iLUVit.event.AlarmEvent;
