+++ src/test/java/FIS/iLUVit/service/ReportServiceTest.java
@@ -9,8 +9,8 @@ import FIS.iLUVit.domain.iluvit.Post;
 import FIS.iLUVit.domain.iluvit.Teacher;
 import FIS.iLUVit.dto.report.ReportRequest;
 import FIS.iLUVit.domain.iluvit.enumtype.ReportType;
-import FIS.iLUVit.domain.iluvit.reports.Report;
-import FIS.iLUVit.domain.iluvit.reports.ReportDetail;
+import FIS.iLUVit.domain.iluvit.Report;
+import FIS.iLUVit.domain.iluvit.ReportDetail;
 import FIS.iLUVit.exception.*;
 import FIS.iLUVit.repository.iluvit.*;
 import org.junit.jupiter.api.BeforeEach;
