+++ src/test/java/FIS/iLUVit/service/TeacherServiceTest.java
@@ -7,8 +7,8 @@ import FIS.iLUVit.dto.teacher.SignupTeacherRequest;
 import FIS.iLUVit.dto.teacher.TeacherDetailRequest;
 import FIS.iLUVit.dto.teacher.TeacherDetailResponse;
 import FIS.iLUVit.dto.teacher.TeacherInfoForAdminDto;
-import FIS.iLUVit.domain.iluvit.alarms.CenterApprovalAcceptedAlarm;
-import FIS.iLUVit.domain.iluvit.alarms.CenterApprovalReceivedAlarm;
+import FIS.iLUVit.domain.iluvit.CenterApprovalAcceptedAlarm;
+import FIS.iLUVit.domain.iluvit.CenterApprovalReceivedAlarm;
 import FIS.iLUVit.domain.iluvit.enumtype.Approval;
 import FIS.iLUVit.domain.iluvit.enumtype.Auth;
 import FIS.iLUVit.domain.iluvit.enumtype.AuthKind;
